+++ src/internal/packager/injector/injector.go
