+++ src/internal/packager/injector/injector_test.go
