+++ src/internal/packager/injector/testdata/expected-injection-pod.json
