+++ src/internal/packager/injector/testdata/expected-injection-service.json
